requirements.txt
python-telegram-bot==20.6
httpx==0.25.0
requests
